refactor: extract authentication logic from git_push & git_pull
