+++ src/main.rs
@@ -15,6 +15,7 @@ mod hook_executor;
 mod rules;
 mod step;
 mod util;
+mod utils;
 mod workflow_queue;
 mod workflows;
